+++ application/client/src/app/service/hotkeys.ts
@@ -22,12 +22,27 @@ import { Action as ExitAction } from '@service/actions/exit';
 import * as Events from '@platform/ipc/event/index';
 import * as Requests from '@platform/ipc/request';
 
+enum BindingState {
+    Bound,
+    Bounding,
+    Unbound,
+    Unbounding,
+    NotInited,
+}
+
 @DependOn(env)
 @SetupService(services['hotkeys'])
 export class Service extends Implementation {
     private _listener!: Listener;
     private _services!: Services;
     private _handlers: Map<string, Map<string, () => void>> = new Map();
+    private _recent: {
+        inputsUpdateState: number;
+        remoteBinding: BindingState;
+    } = {
+        inputsUpdateState: -1,
+        remoteBinding: BindingState.NotInited,
+    };
 
     public override ready(): Promise<void> {
         const channel = ilc.channel(this.getName(), this.log());
@@ -191,7 +206,9 @@ export class Service extends Implementation {
             inputs: (): void => {
                 // We have to use here litle delay, because some angular material components makes changes
                 // asynch. To catch last state of components we have to let "them" update itselfs
-                setTimeout(() => {
+                clearTimeout(this._recent.inputsUpdateState);
+                this._recent.inputsUpdateState = setTimeout(() => {
+                    this._recent.inputsUpdateState = -1;
                     if (document.activeElement === null) {
                         return;
                     }
@@ -203,7 +220,7 @@ export class Service extends Implementation {
                         this._listener.requirement(Requirement.NoInput).activate();
                         this.remote().bind();
                     }
-                }, 150);
+                }, 150) as unknown as number;
             },
             session: (): void => {
                 const active = this._services.system.session.active().session();
@@ -226,6 +243,12 @@ export class Service extends Implementation {
     } {
         return {
             bind: (): void => {
+                if (
+                    [BindingState.Bound, BindingState.Bounding].includes(this._recent.remoteBinding)
+                ) {
+                    return;
+                }
+                this._recent.remoteBinding = BindingState.Bounding;
                 Requests.IpcRequest.send<Requests.Hotkey.On.Response>(
                     Requests.Hotkey.On.Response,
                     new Requests.Hotkey.On.Request(),
@@ -235,15 +258,27 @@ export class Service extends Implementation {
                             this.log().error(
                                 `Fail to activate listener of hotkeys: ${response.error}`,
                             );
+                            this._recent.remoteBinding = BindingState.NotInited;
+                        } else {
+                            this._recent.remoteBinding = BindingState.Bound;
                         }
                     })
                     .catch((err: Error) => {
                         this.log().error(
                             `Fail send request to activate listener of hotkeys: ${err.message}`,
                         );
+                        this._recent.remoteBinding = BindingState.NotInited;
                     });
             },
             unbind: (): void => {
+                if (
+                    [BindingState.Unbound, BindingState.Unbounding].includes(
+                        this._recent.remoteBinding,
+                    )
+                ) {
+                    return;
+                }
+                this._recent.remoteBinding = BindingState.Unbounding;
                 Requests.IpcRequest.send<Requests.Hotkey.Off.Response>(
                     Requests.Hotkey.Off.Response,
                     new Requests.Hotkey.Off.Request(),
@@ -253,12 +288,16 @@ export class Service extends Implementation {
                             this.log().error(
                                 `Fail to deactivate listener of hotkeys: ${response.error}`,
                             );
+                            this._recent.remoteBinding = BindingState.NotInited;
+                        } else {
+                            this._recent.remoteBinding = BindingState.Unbound;
                         }
                     })
                     .catch((err: Error) => {
                         this.log().error(
                             `Fail send request to deactivate listener of hotkeys: ${err.message}`,
                         );
+                        this._recent.remoteBinding = BindingState.NotInited;
                     });
             },
         };