+++ scripts/tools/os.rb
@@ -18,6 +18,12 @@ module OS
     OS.unix? && !OS.mac?
   end
 
+  def self.arm64?
+    arch = (OS.unix? || OS.mac?) ? `uname -m` : `echo %PROCESSOR_ARCHITECTURE%`
+    arch.chomp!.downcase!
+    arch=='arm64' || arch=='aarch64'
+  end
+
   def self.executable(filename)
     exe = if OS.windows?
             '.exe'